Détécter les changements rapide de vitesse longitudinales
Changement apportés :
- Utilisations du timestamp fournit par la position GPS en lieu et place du timestamps fournit par le système Android. L’utilisation de ce timestamp permet de calculer avec plus de précisions le temps qui s’est écoulé entre la réception de deux points GPS.
- Mise en place d’un filtrage médian qui remplace le filtrage moyen. Ce filtre à pour effet de gommer les valeurs erratiques discrètes.
- Ne plus exclure les points GPS pour lesquels la vitesse est inférieur à 20 km/h.  Ces points sont très important pour la détection de freinage. Le fait d’exclure ces points avait pour effets de ne pas détecter les freinages pour lesquels la vitesse du véhicule devenait inférieur à 20 km/h. C’est le cas, par exemple, lorsque le véhicule freine brusquement pour s'arrêter.

+++ BoxPreventium/app/src/main/java/com/preventium/boxpreventium/server/EPC/ForceSeuil.java
@@ -16,13 +16,13 @@ public class ForceSeuil {
     public int index = -1;
     public short IDAlert = -1; // IDAlert.
     public short TPS = 0; // Temps de déclenchemnt en secondes.
-    public double mG_low = 0.0; // Seuil bas en mG.
-    public double mG_high = 0.0; // Seuil haut en mG.
+    public long mG_low = 0; // Seuil bas en mG.
+    public long mG_high = 0; // Seuil haut en mG.
     public LEVEL_t level = LEVEL_t.LEVEL_UNKNOW;
     public FORCE_t type = FORCE_t.UNKNOW;
 
     ForceSeuil(){};
-    public ForceSeuil(int index, short IDAlert, short secs, double mG_min, double mG_max) {
+    public ForceSeuil(int index, short IDAlert, short secs, long mG_min, long mG_max) {
         this.index = index;
         this.IDAlert = IDAlert;
         this.TPS = secs;
